pipeline/requirements.txt
beautifulsoup4==4.12.3
httpx==0.27.2
openai==1.45.0
pdfplumber==0.11.4
python-dotenv==1.0.1
requests==2.32.3
supabase==2.7.4
